services/sentiment/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
transformers>=4.40.0
torch>=2.0.0
hf_xet
finnhub-python
python-dotenv
